Better column type determination logic for JDBC #3878
Summary:
The approach earlier was to infer the field names, and their types from the first record in the result set.
This is brittle because we may either miss a field, or misinterpret the type as null if the field was null in the first record.
With this approach we now iterate over all the rows in the result set to get the fields headers and the corresponding non-null types, if any.
Note that having mixed types would still throw an error.

Test Plan: Tested with sparse fields and mixed types in Tableau and DataGrip

Reviewers: kw, dhruba, ari, sandeep

Reviewed By: ari

Differential Revision: https://rockset.phacility.com/D3536