Update ErrorModel swagger to fix ErrorModel enum deserialization (#38)
Updates the code generated file for ErrorModel.java. This adds a @JsonCreator annotation on the toString method, which allows the values to be correctly deserialized from the enum value instead of the name.

This annotation is added onto many other enums, but they have not been added in this PR because there are multiple breaking changes. This solution replaces https://github.com/rockset/swagger-codegen/pull/2 and addresses https://github.com/rockset/rockset-java-client/issues/37