Patch time parsing errors. (#41)
We were seeing errors in our logs such as `Invalid time from server: 20:06:01.434197`.

After some experimentation, we observed that the JDBC driver just calls `java.sql.Time.valueOf(String)` was being called. This method explicitly expects a string of the format:

```
     * @param s time in format "hh:mm:ss"
```

When running a few cast queries through the rockset console, we observed that:

```
    // from console:
    // 1 select cast('12:34' as time)    -> __rockset_type= 'TIME' / value = '12:34:00'
    // 2 select cast('12:34:56' as time) -> __rockset_type= 'TIME' / value = '12:34:56'
    // 3 select cast('12:34:56.1' as time) -> __rockset_type= 'TIME' / value = '12:34:56.1'
    // 4 select cast('12:34:56.123' as time) -> __rockset_type= 'TIME' / value = '12:34:56.123'
    // 5 select cast('12:34:56.123456' as time) -> __rockset_type= 'TIME' / value = '12:34:56.123456'
    // 6 select cast('12:34:56.1234567890' as time) -> __rockset_type= 'TIME' / value = '12:34:56.123456'
```

Some of these will cause exceptions to bubble up from the JDBC driver.

This patch fixes the parse exceptions. I stuck with Joda vs. porting to java.time/JEP150 because I was not sure of the knock-on effort of having to move everything else over.