Move buildRocksetObject call to addCall (from add).
We need to peek into the response stream so that we can pull the last_offset and wait on the fencing / we need read-after-write functionality.

We've built a shim to do this, but we also noticed that we lost JEP150/java.time type mapping into Rockset. This was because the plumbing that does this in the Java SDK happened in the add() method and we are using addCall().

It seems that add() will, eventually, call addCall() downstream.